refactor: simplify onboarding logic
